pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "lirmotiffinder"
version = "0.1.0"
description = "Find LIR motifs in intrinsically disordered regions predicted by metapredict"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["bioinformatics", "autophagy", "LIR", "LC3", "IDR", "metapredict", "disordered"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "biopython>=1.79",
    "metapredict>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-cov",
]

[project.scripts]
lirmotiffinder = "lirmotiffinder.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/lirmotiffinder"
"Bug Tracker" = "https://github.com/yourusername/lirmotiffinder/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["lirmotiffinder*"]
